[feat] wrap io in separate package
